PaymentInfo as interface #1053
